Fix #3: prevent server crash when upgrading uncoloured shulker (#4)
`getColor` is client-only Dist hence the side violation that causes a
crash. I'm not sure why all of this information needs to be client only.

The only instance is the actual shulker box itself, which is not
client-only (as it's actually in the block instance). Considering the
block has been placed, this is the safest way to get the colour.

Similarly, removed the @DistOnly in your own class as there's no need
for it to be like that, meaning that the `getColor` function can be used
in the other side of the if block with the iron shulker box tile entity.

Finally, if the colour returned is `null` (meaning that it's an
uncoloured shuler box), I've set the default to purple -- which matches
the same colour of the upgraded shulker boxes (although not the colour
of the *actual* purple shulker box).