Update zh_cn.json
"Hunt and Gather" does not have an official translation.
"Bonfire Lit" is from "Dark Soul" iirc. Considering that its Chinese translation is not from the game itself, I decided not to translate it.
"Fire Up the Grill" is a little bit difficult to translate because there is no similar phrase of it in Chinese as far as I know, so I will leave it here to see if there is someone who is able to translate it without losing its own meaning.