Fix Modrinth README badge
API v1 has been deprecated for over a year now and should no longer be used. There are two options for implementation of this fix:
- Use the dedicated shields.io route
- Update it to still manually construct the request with API v2

Each has advantages and disadvantages. Namely, the former no longer has the precise count, and the latter will need manual reconstruction every time that the API updates. Considering that this is down to the clock with how many deprecation warnings have gone out at this point, I believe the former is the best option, and thus the one I have implemented in this PR. If you believe the latter is better, feel free to request that change.