Add localized name for the Rotten Tomato Entity
The Rotten Tomato Entity localized name is shown when a Rotten
Tomato flies and is pointed by player cursor with mods like Waila.

This patch adds the missing localized name for the Rotten Tomato
entity, to every languages with a Rotten Tomato Item translation key,
as it reuses the same translation for the Entity.