Always clamp current power when changing max power (#7443)
If the new max power was lower than the current power, the clamping took
only place if the new max power was larger than the default max power,
i.e. it worked for any energy card but the last one.

Fixes AppliedEnergistics#7431

There is a potential for refactoring here, depending on whether the 3
functions setAEMaxPowerMultiplier, setAEMaxPower, and resetAEMaxPower
are considered to be a stable API or not. I chose to not refactor for
compatibility reasons. Deciding if that is the way to go is up to a
reviewer in my opinion.

Be advised that this is my first time coding for any Minecraft mod, but
the change should be small enough for a starter task.