Improved ME Terminal Search to be more similar to REI/EMI (#7883)
This was created based on #7311, instead of relying on the functionality
of JEI, REI, etc. I modified the search functionality to search similar
to RS, JEI, etc.

The reason behind this PR is the same reason as the previous PR (#7311):
>This can be useful because it is annoying when you try to search
something with the modid and the name like it is possible in JEI. For
example "@ae2 crystal". Without it, it tries to search a mod with the
name "ae2 crystal".

---------

Co-authored-by: Sebastian Hartte <shartte@users.noreply.github.com>