Avoid ArrayIndexOutOfBoundsException when max health/hunger is changing
If the max health/hunger becomes larger between the generation of the offsets and the rendering of the bar, then it could try to index into the offsets Vector using an index that is larger than the size of the Vector. After this commit, we avoid that possibility and use a default offset of 0,0 for icons past the end of our calculated offsets.

Fixes #308