Prevent entries added through CopperRegistries from crashing the game
Most reported issues of this happening are actually caused by a
different issue during registration that is suppressed due to the crash.
Now that the crash can't happen, the underlying exception will be
properly reported. This change also allows using Create with the
Compressed mod, but if Compressed loads before Create, Create's
additions to the CopperRegistries will fail to apply.