Fix race in Contraption simplifiedEntityColliders update
This issue is most obvious in elevators with multiple doors. When the
collider gets invalidated multiple times in quick succession, multiple
futures are created. If a future which was started before all door state
updates were finalized completed after all other futures, the final
simplifiedEntityColliders would be outdated and invalid (containing
ghost doors, or in the case of an elevator in motion, missing door
collisions).

This commit addresses this issue in two ways:

First by cancelling any existing future to ensure that any future which
completes was started after the most recent invalidation.

Second by removing the null assignment of
simplifiedEntityColliderProvider from the future. This prevents the
future from becoming null after a null check and before a cancellation
(the only time where the null value matters). Cancelling a future twice
is not an issue so there's no need to track if the future is null other
than to avoid a null dereference.