Allow filtering config stack slots with containers holding 0 of a stack (#8324)
This PR allows config slots of `CONFIG_STACKS` mode to be filtered with
a minimum of 1 of a given stack retrieved from a container, even if the
container may have a zero amount of that stack.

The rationale for this comes from
https://github.com/62832/ArsEnergistique/issues/17 and the [resulting
Mixin](https://github.com/62832/ArsEnergistique/blob/master/src/main/java/gripe/_90/arseng/mixin/ConfigInventoryMixin.java)
used to allow players to filter ME Interface slots to Source, even if
the Source Jar being used to filter it is empty. As the Ars Énergistique
add-on is designed around a key type for which only one possible
resource exists (Source), it is wholly unambigious that any Source Jar
may contain or make use of Source regardless of whether it currently
doesn't contain any of it.

The same would apply, should it ever be updated to 1.21, to the Applied
Botanics add-on and empty Mana Tablets as a result of this being
included upstream, and has already applied as such, though only also in
the presence of Ars Énergistique.